first-langgraph-app/pyproject.toml
[project]
name = "first-langgraph-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.7",
    "langchain-azure-ai>=1.0.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langsmith>=0.6.7",
    "notebook>=7.5.3",
    "python-dotenv>=1.2.1",
]
